LS-19174 Unsigned ints hack (#2)
* copy paste these over from josh's fork

https://github.com/ripjar/cld2/commit/79be1adea78f0d376cb793f4dae8e70b100dadcc

this commit is not on a branch, so make it so

* bump version

* bump version

+++ cld2/internal/scoreonescriptspan.cc
@@ -22,7 +22,6 @@
 #include "cldutil.h"
 #include "debug.h"
 #include "lang_script.h"
-#include <stdint.h>
 
 #include <stdio.h>
 
@@ -185,7 +184,7 @@ void GetTextSpanOffsets(const ScoringHitBuffer* hitbuffer,
 
 
 int DiffScore(const CLD2TableSummary* obj, int indirect,
-              uint8 lang1, uint8 lang2) {
+              uint16 lang1, uint16 lang2) {
   if (indirect < static_cast<int>(obj->kCLDTableSizeOne)) {
     // Up to three languages at indirect
     uint32 langprob = obj->kCLDTableInd[indirect];
@@ -379,7 +378,7 @@ uint16 NextChunkLang(const SummaryBuffer* summarybuffer, int i) {
 //
 // We go out of our way to minimize the variation in the ResultChunkVector,
 // so that the caller has fewer but more meaningful spans in different
-// languages, for the likely purpose of translation or spell-check.
+// lanaguges, for the likely purpose of translation or spell-check.
 //
 // The language of each chunk is lang1, but it might be unreliable for
 // either of two reasons: its score is relatively too close to the score of
@@ -437,7 +436,6 @@ void SummaryBufferToVector(ScriptScanner* scanner, const char* text,
       if (n >= n_limit) {n = 0;} // New boundary not found within range
 
       // Also back up exactly one leading punctuation character if '"#@
-      // 'random', "quotes", #hashtags, @handles
       if (n < n_limit) {
         unsigned char c = us[-n - 1];
         if ((c == '\'') || (c == '"') || (c == '#') || (c == '@')) {++n;}
@@ -671,7 +669,7 @@ void DumpSummaryBuffer(FILE* df, const SummaryBuffer* summarybuffer) {
 int BetterBoundary(const char* text,
                    ScoringHitBuffer* hitbuffer,
                    ScoringContext* scoringcontext,
-                   uint8 pslang0, uint8 pslang1,
+                   uint16 pslang0, uint16 pslang1,
                    int linear0, int linear1, int linear2) {
   // Degenerate case, no change
   if ((linear2 - linear0) <= 8) {return linear1;}
@@ -1144,8 +1142,8 @@ void ScoreEntireScriptSpan(const LangSpan& scriptspan,
   if (scoringcontext->flags_cld2_html) {
     ChunkSummary chunksummary = {
       1, 0,
-      one_one_lang, UNKNOWN_LANGUAGE, score, 1,
-      bytes, 0, scriptspan.ulscript, reliability, reliability
+      static_cast<uint8>(one_one_lang), UNKNOWN_LANGUAGE, static_cast<uint8>(score), 1,
+      static_cast<uint8>(bytes), 0, static_cast<uint8>(scriptspan.ulscript), static_cast<uint8>(reliability), static_cast<uint8>(reliability)
     };
     CLD2_Debug(scriptspan.text, 1, scriptspan.text_bytes,
                false, false, NULL,
@@ -1333,4 +1331,3 @@ void ScoreOneScriptSpan(const LangSpan& scriptspan,
 }
 
 }       // End namespace CLD2
-
